stock_producer/requirements.txt
kafka-python==2.0.2
yfinance==0.2.28
schedule==1.2.0
python-dotenv==1.0.0
pandas>=1.3.0
